requirements.txt
pyyaml
deepdiff
rich
